# Improve performance of `StreamSelectLoop` when no timers are scheduled